Provide some sanity checking to the codebase

+++ cpp-commons/headers/io/CSVFile.hpp
@@ -7,21 +7,23 @@
 #include <sstream>
 #include <vector>
 
+#include "utils/Sanity.hpp"
+
 class CSVFile {
 
 public:
     CSVFile(std::istream&, char);
 
-    std::string get(const std::string&, unsigned) const;
+    std::string get(const std::string&, long) const;
     std::map<std::string, std::string> find(const std::string&, const std::string&) const;
-    unsigned size() const { return data.size(); }
-    unsigned columnCount() const { return columns.size(); }
+    unsigned long size() const { return data.size(); }
+    unsigned long columnCount() const { return columns.size(); }
 
 private:
     std::vector<std::vector<std::string>> data;
-    std::map<std::string, unsigned> columns;
+    std::map<std::string, long> columns;
 
-    void parseHeader(std::istream&, unsigned, char);
+    void parseHeader(std::istream&, unsigned long, char);
     void parseData(std::istream&, char);
     unsigned getColumn(const std::string&) const;
